Move UserMedia to MediaDevicesShim
